Refactor. 소스(파일) 줄이기
